+++ src/enums/build_runtime.rs
@@ -65,12 +65,6 @@ pub enum BuildRuntime {
     PythonMl312,
     #[serde(rename = "python-ml-3.13")]
     PythonMl313,
-    #[serde(rename = "deno-1.21")]
-    Deno121,
-    #[serde(rename = "deno-1.24")]
-    Deno124,
-    #[serde(rename = "deno-1.35")]
-    Deno135,
     #[serde(rename = "deno-1.40")]
     Deno140,
     #[serde(rename = "deno-1.46")]
@@ -224,9 +218,6 @@ impl BuildRuntime {
             BuildRuntime::PythonMl311 => "python-ml-3.11",
             BuildRuntime::PythonMl312 => "python-ml-3.12",
             BuildRuntime::PythonMl313 => "python-ml-3.13",
-            BuildRuntime::Deno121 => "deno-1.21",
-            BuildRuntime::Deno124 => "deno-1.24",
-            BuildRuntime::Deno135 => "deno-1.35",
             BuildRuntime::Deno140 => "deno-1.40",
             BuildRuntime::Deno146 => "deno-1.46",
             BuildRuntime::Deno20 => "deno-2.0",